+++ Demo/Module2/src/main/java/com/xiaojinzi/component2/view/Component2Act.kt
@@ -151,7 +151,7 @@ class Component2Act : BaseAct() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.component2_act)
         Component.inject(this);
-        tv_data.text = data;
+        tv_data.text = data
 
         val fragment = Router
                 .with(ModuleConfig.Module1.TEST_FRAGMENT)